Always resolve icons using the default screen
When a custom icon theme path is specified by the indicator,
we create a new GtkIconTheme which additionally contains this
path and use that to resolve the icon name. We unfortunately forgot
that without setting the screen, this new theme object can't know the
current icon theme (at least not under X). This leads to icons names being
resolved from the "hicolor" theme only, which is especially ugly if the
current theme has nice symbolic icons.

Fixes #32