Reimplement the icon actor without fancy vfuncs
This solves a nasty assert during garbage collection.

<rant>
Looking at the mess, I want to puke. Doing JavaScript in Gnome means
using not one, not two but three(!) methods of garbage collection. There
is the JavaScript garbage collector, the GObject reference counting and
the destroy() methods. As soon as you mess with GObjects (like, subclass
one, override a vfunc or connect a signal) you're in for creating
dangerous reference cycles between C and JS (i.e. a leak). In order to
fix it, we use destroy() methods. So we're actually doing manual memory
management, but worse. The object is not gone but lingers around
like a zombie and can still leak if we forgot to clean up a cycle. Or if
someone forgets to call the destroy() method. I thought JavaScript was
such a great language because we don't have to think about memory since
the GC will always save us. So great that a debugger or a memory tool
is not needed for Gjs at all. Deciphering a stack trace in the log file
is the way to go. Reasoning about when which parts of an object
are going to be freed (if at all) is driving me insane.

Why do I have to debug asserts in C code that runs during the garbage
collection phase? Why can I provoke C asserts using semmingly good
javascript code? Fuck Javascript. Fuck Gjs. And fuck garbage collection.

I'm looking forward to Wayland, where a Shell crash will take down my
whole session. That's when the fun in debugging begins.
</rant>