Don't add "-panel" to the end of the icon_name unless icon_name is not in the current theme
Always adding "-panel" to the end of icon_name assumes that icon_name + "-panel" exists or should exist. That is not always the case.

This commit 1st checks to see if icon_name +  "-panel" is in the current icon theme and if it is fine use it, otherwise it will use icon_name as is.