Add icon opacity, size, position and other effects settings
The icons can be very colorful. Some people might not like it and older
icons do not blend too well with newer monochrome ones.

Add settings to configure multiple values of the icons and in
particular:
 - Icons position in the panel
 - Icon size
 - Opacity
 - Desaturation
 - Brightness
 - Contrast

As per this a new preference dialog is provided to configure such values

Closes #196
Fixes #142, #215