Add desaturation (hard coded)
[why]
The icons can be very colorful. Some people might not like it and older
icons do not blend too well with newer monochrome ones.

[how]
Just fiddle the code of TopIconsPlus in.
The numbers are hard coded and just how I like it.

Copying over the settings dialog would not too hard, too.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>