appIndicator: Don't try to reload multiple times a pending icon
In case we're reloading an icon for which we've a pending operation in
progress we were just canceling the operation and request it again.
However this proves to be more CPU intense on some indicators that update
the icon quite often (hello again indicator-multiload!), so given the
previous operation for the same icon didn't finish yet we can just wait for
that instead of going through async calls cancellation.