Util: Cancel previous properties refresh requests on new one
When multiple calls to refreshPropertyOnProxy() are done for a single proxy
we may end up emitting multiple times properties changed events, while we
care only about the last one.

So, use cancellables and keep track of them in order to stop a proxy call
if a new one has just been requested.

Also stop any ongoing request when an indicator is closed.