appIndicator: Settings: Fix crash on click-and-hold
[why]
A crash occurs when in the settings dialog the + or - buttons of opacity
are clicked and held for a prolonged time (e.g. for almost a complete
change from 0 - 200 or equally).

This has not been observed with the other settings. But they have a far
smaller range, resulting in shorter hold durations.

[how]
While pushed we constantly create new Gio.Settings objects. This seems to
lead to some kind of overload, memory shortage or number of handles
limit? I don't know. (And why would we want to do it anyhow.)

Instead we should create one Gio.Settings object and reuse it
everywhere. In this way we will not run out of handles/memory/whatever.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>