appIndicator: Maximize opacity on hover
[why]
The system icons in the title bar light up when the cursor hovers over
them. Although we can now adjust (dim) the opacity of our icons, this
behavior is still missing.

[how]
Set opacity to maximum on hover.
Reset to selected value when cursor leaves the icon.

[note]
The effect is not as pronounced as the system icons it seems. Maybe they
also increase the brightness or just use brightness for the effect. But
the change introduced with this commit seems ok.
One could think about also using original brightness, contrast, ...
whatnot, but I reckon that will be heavily depending on personal taste.
Making this configurable is a possible way, but this is too much for my
taste.
At least on my setup this solution is quite ok, and it is always a
compromise anyhow. My opacity setting is 200.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>