appIndicator: Add possibility to choose icon size
[why]
People seem to have different setups and different opinions on how big
the icons should be. The system icons size also differs depending on
theme - I have heard.

[how]
Let the user choose a pixel size. The value zero (0) means the handling
of this extension is unchanged and the size is automatically choosen.

[note]
Obviously the settings callback was still in the codebase; I must have
forgotten to remove it when the old icon size code has been removed.