appIndicator: Fix 'reduce padding' for some icons
[why]
We remove all style settings from the gnome-shell.css that usually
affect the menu buttons. This includes the default icon size.

When the icon has some 'other' size, this will leave us with garbage.

[how]
Just overwrite the padding of the system gnome theme. Keep all other
style settings intact (including icon size).

[note]
We do not query the current style to build the new style, because (at
least at the moment) no other style setting is done anywhere. When in
some point in the future a style is set before this new line (which in
unlikely, because it is at the top of the constructor), we would have to
assemble the new setting more carefully.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>