appIndicator: Fix resizing pixmap icons
[why]
Icons created from pixmaps can not be resized (properly).

[how]
When we create a icon from a pixmap we actually crate the icon in the
native size. For example if the smallest available pixmap is 22 px the
icon will also be 22 px, regardless of requested icon size.

This is ok if the pixmap is by chance the requested size. But this is
not always the case, especiually if the icon size is user selectable.

So after creating the pixmap image we tell it how big it shall be.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>