appIndicator: Fix pixmap icons sizes
[why]
When we update icons we first create the normal (nominal 16 px) icons
and afterwards the overlay ones (nominal 10 px).

If the application has just pixmap icons we succeed for both types.
Because the overlay icon will be queried last, this is the one that will
be used. This can be too small.

[how]
Update the overlay icon before the normal icon. If we have just one
pixmap icon we will end up with a correctly sized one.

[note]
This is a bit unclear for me, how the handling of overlay icons is
supposed to work. I have no such application to test that (suggestions
welcome).
Maybe this is (only) an issue with the icon cache. The pixmaps icons are
not cached, while normal ones are. Maybe a better fix would be to also
cache the overlay icons. Hmm, but this is a bigger change than I'm
willing to propose within a PR with another goal althogether.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>