Wait until Gtk.IconTheme.get_default() works
It seems that the reason for not showing sometimes the appindicators
during startup is because Gtk.IconTheme.get_default() returns NULL.
This seems to happen only very early during startup, so this patch
waits until it returns a valid value before continuing enabling
the extension.

Of course, this patch is just a hack. If this really fixes the
problem, the bug must be fixed upstream.