appIndicator: Ignore further icons creation during an idle
And also remove the idle if instead we got a valid icon meanwhile.

This may happen in the case that the icon ID (and so name or size) changed
since the time we requested an invalid icon, and in such case we would still
have the idle running.

So we need to ignore a new one in case the new icon is still invalid, while
we need remove the source if the new icon is available.