iconCache: Rewrite simplifying the usage for GIcon's only
The icon cache has been built with the idea of caching actors, however now
we only cache the actual actor contents which is eventually always a GIcon,
so there's no point to complicate it.

Also use native JS Maps and remove destroy notification support, as icons
have not such capability anymore.

As per this, when returning the GIcon added to the cache we need to make
sure that we're returning the object that is actually in the cache and not
an equal object that may be already there.