util: Try to batch properties updates when they comes close enough
Use a quick timeout (set to 16ms, that is still our 60fps target) to make
possible to batch properties changes so that if more than one come together
we can just send out one signal.

This also ensures that many inter-connected properties are handled in order.

Potentially we could also stop the property changes emission when a property
request is cancelled, but if we already got a result when another one is
requested we can proceed sending the one we already have, without need to
complicate the code here.