statusNotifierWatcher: Use a simple promise to wait startup
As per promise abilities, we can now just await for startup completion
and once there, create the visual elements and emit signals about the
registered items.

There's no need to check if the item has been removed, as if this happens
the item cancellable will be cancelled, and thus the async function will
be stopped.