promiseUtils: Disconnect cancellable when the promise is resolved
In CancellablePromise we were connecting to the cancellable in order to
monitor its cancellation, however we never disconnected from it and this
was causing gjs to keep a reference on the promise itself, being the
scope of the cancellable callback, and thus never garbaging collecting
both the promise and its children (such as GSources).

So, connect earlier to the signal and once we've resolved the promise,
disconnect form the cancellable.

Helps with #277