promiseUtils: Disconnect cancellation signal after resolving or rejecting
As per the previous commit we do disconnect from cancellation, but this
happens too early so the cancellation event isn't ever tracked.

We need in fact to handle it after that the promise is resolved or
rejected.

In the rejected case, we need to be extra careful though, as in case
the cancellation happens as per cancellation's ::cancelled signal
callback we need to do this in the next idle cycle or we'll deadlock.