appIndicator: Use custom theme icon based on Shell one as default theme
Especially under wayland we should avoid to get the theme from the
default gtk one as it may be always unset, as it depends on the Gdk
Screen that is undefined in such scenario.

So add an helper function to create a new Gtk theme using the shell's
gtk icon theme and try to use it when no default gtk theme is available.

We could probably do this in all the cases, but for now let's keep X11
environment as it is.

See: https://gitlab.gnome.org/GNOME/mutter/-/issues/960

Closes #279