indicatorStatusIcon: Only redirect events coming from tray icon container
It seems that processing events that we get from the tray icon itself
results in mutter not accepting them when running in pure wayland, as
the clients should be in an "allowed-list" of windows who can require
grabs.

If we do this from a shell actor, instead, this works. So let's just
use the container events.

Fixes: #301