Add indicator for recent icon names in prefs dialog
[why]
When someone wants to exchange an icon with a custom one the icon name
is needed. That can be nontrivial to find it out.

[how]
Just collect all icon names that we encountered since the extension has
been started and list these names below the icon replacement table in
the settings dialog.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>