appIndicator: Notify property changes promptly if content is provided
For some properties (Status and Icons theme path) we get the changed
value as the status argument, so in such case we don't have to perform
further dbus calls but we can just emit the change promptly, or wait
for next update if one is already pending.

This is also fixing a nasty issue we had that was causing icons state
not being updated when they were moving to passive just before removing
the indicator from the bus.

This is happening in electron apps that provide an option to toggle the
indicator (such as ferdium): when turning it off the old indicator sends
us a status change to passive and the object is immediately removed from
the bus. This was problematic for us because we were trying to update
the state again from an object that had been removed, leading to dbus
errors.

Closes: #348