appIndicator: Do not allow overlay icons to interfere loading of others
If an overlay icon is loading it must not cancel any operation for other
icons, instead we should just proceed or we'd block the loading of all
the icons.

It's not worth adding another list to keep track of icons by their type
because, we want loading any icon to stop the previously started
operations, whether they're attention or normal ones, while overlay ones
are so uncommon that we don't really care much if we require updating
them multiple times.