appIndicator: Fix logic to get the name owner when using name watcher
In KDE applications we use the name watcher to monitor applications
indicator presency, however we did not consider that in our logic
properly.

This fixes the case in which a Qt applications (i.e. Telegram) hides an
indicator from settings. Before of this we used not to remove it, even
though was not working anymore.