appIndicator: Support indicator-multiload again
While it was showing, St.Icon is meant to show icons with the same width
and height, so even changing the size we'd end up in an icon with the
wrong aspect ratio.

As per this, use native ClutterActor with content generated by
St.TextureCache as a child, in this way we can paint it as expected.

The downside here is that we need to create a special paint vfunc,
hoping this won't cause much overhead for the other icons.

LP: #1988791