appIndicator: Do not override paint vfunc unless in specific cases
We may need to use a special class with overridden paint function only
in the few cases in which the indicator requires a custom image (so
basically we know only one, even if it's popular: indicator-multiload).

To avoid making JS not to overdo the job that should stay in the C-land,
implement an IconActor subclass with the said vfunc (only, to avoid
doing the copy and paste job here), and in case we require to paint a
custom image, emit a signal that makes us re-create the icon actor with
the required vfunc.

It's not super-nice, but it would definitely avoid to make all the other
indicators to be potentially limited by something they don't need.