statusNotifierWatcher: Do not try to use a destroyed indicator
In the name-owner-changed signal handler we have we use an async
function to check if the indicator is still alive, however by the moment
we're finished the invocation, the indicator may have been already
destroyed, and so we should not try to access to it.

In theory we could handle this also by using the indicator cancellable
instead but it's better not to expose internal logic too much here.