appIndicator: Never deep unpack pixmaps properties
Pixmaps icons properties may be huge arrays of arrays of bytes, unpacking
them could be expensive and most likely unneeded.

We did this in the past as it made easy to handle the properties, but it's
time to change this, so always use the variants as they are extracting the
child values of the arrays to perform size comparisons, until we actually
need the bytes to perform the RGBA conversion.