appIndicator: Add support for off-process loading of pixmaps
Loading pixmaps can be a quite slow sync operation, due to the
rgba-conversion we need to do ourself, and this can slowdown some scenarios
such as at shell unlock. Gjs does not support (yet) JS workers, but we can
still simulate such behavior by using an external process that we handle
fully asynchronously to get a byte array that can eventually be just passed
to GdkPixmap to fill the bytes.

While this can be slower in terms of time because of the overhead and the
double dbus calls, it will still be be beneficial because it won't cause any
rendering slowdown to the shell process. That was visible especially during
lock/unlock operations.

Helps with: #295