util: Use an async generator to iterate over bus objects
We need to do chunked requests to avoid initialization slowdowns, and
apparently awaiting many promises here is not as good as expected and may
lead to hanging, while using a generator is better, so  let's do this.

This implies bumping the required eslint version, while the the minimum
shell version should be ok, because gjs in GNOME 3.34 required mozjs60,
while async generators are supported since mozjs55.

Helps with: #295