dbusMenu: handle LayoutUpdated signal for empty menus
This problem is caused by the "empty menus":

For a LayoutUpdate signal, class DBusClient only requests a layout update
when this._active is true (indicating the menu is currently open), or it
is delayed until this._active becomes true (when the menu is opened later).
this._active is only set to true when an open-state-changed signal is
emitted, which is triggered by opening the popup menu.
But an empty menu will never open, so this._active keeps false, so the
delayed LayoutUpdate handler never runs.

QQ is an IM app who provides no menu entries before user logged in,
and updates it when logged in. But it probably does not emit a LayoutUpdated
signal before updating menu so its tray icon menu can never open until
appindicator restarts (by disabling the extension or ending gnome session).

Here we try updating entries before opening an empty menu.
Only primary click and secondary click will trigger an update.

Co-Authored-By: Matt Mower <mowerm@gmail.com>
Co-Authored-By: Marco Trevisan <mail@3v1n0.net>

Closes: #494